// frontend/src/api/models/handlerCategoryTreeNode.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Category tree node with children
 */
export interface HandlerCategoryTreeNode {
  children?: HandlerCategoryTreeNode[]
  code?: string
  description?: string
  id?: string
  level?: number
  name?: string
  parent_id?: string
  sort_order?: number
  status?: string
}