// frontend/src/api/models/handlerCreateCategoryRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Request body for creating a new category
 */
export interface HandlerCreateCategoryRequest {
  /**
   * @minLength 1
   * @maxLength 50
   */
  code: string
  /** @maxLength 2000 */
  description?: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string
  parent_id?: string
  sort_order?: number
}