// frontend/src/api/models/handlerUpdateCategoryRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

/**
 * Request body for updating a category
 */
export interface HandlerUpdateCategoryRequest {
  /** @maxLength 2000 */
  description?: string
  /**
   * @minLength 1
   * @maxLength 100
   */
  name?: string
  sort_order?: number
}